+++ Resources/sql/MigrateFosUser.sql
@@ -11,6 +11,7 @@
  * - create a `user_migrate` table,
  * - modify all the needed fields,
  * - add missing ones.
+ *
  * Then, when you are ready, you can:
  * - rename your FOSUSerBundle table to `user_fosuserbundle` (or whatever you want),
  * - rename the `user_migrate` one to `user`.
@@ -34,34 +35,33 @@ ALTER TABLE user_migrate DROP COLUMN username_canonical;
 ALTER TABLE user_migrate DROP COLUMN email_canonical;
 ALTER TABLE user_migrate DROP COLUMN confirmation_token;
 
+-- Add identifier field
+ALTER TABLE user_migrate ADD identifier varchar(32) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER id;
+-- Fills in with unique data
+UPDATE user_migrate SET identifier = MD5(CONCAT(email, id));
+-- Make field as NOT NULL
+ALTER TABLE user_migrate MODIFY COLUMN identifier varchar(32) NOT NULL;
+
 -- Add indexes
 ALTER TABLE user_migrate ADD CONSTRAINT PRIMARY KEY (id) USING BTREE;
 ALTER TABLE user_migrate ADD CONSTRAINT un_email UNIQUE KEY (email);
+ALTER TABLE user_migrate ADD CONSTRAINT un_identifier UNIQUE KEY (identifier);
 
--- Modify fields + add token one
+-- Modify fields
 ALTER TABLE user_migrate MODIFY COLUMN id bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT;
-ALTER TABLE user_migrate MODIFY COLUMN email varchar(128) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL;
-ALTER TABLE user_migrate CHANGE last_login latest_signin datetime NULL;
+ALTER TABLE user_migrate CHANGE email email varchar(128) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL AFTER identifier;
+ALTER TABLE user_migrate CHANGE creation creation datetime DEFAULT NULL AFTER email;
+ALTER TABLE user_migrate CHANGE enabled enabled tinyint(1) DEFAULT 0 AFTER creation;
+ALTER TABLE user_migrate CHANGE salt salt varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER enabled;
+ALTER TABLE user_migrate CHANGE `password` `password` varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL AFTER salt;
+ALTER TABLE user_migrate CHANGE last_login latest_signin datetime NULL AFTER `password`;
 ALTER TABLE user_migrate ADD token varchar(40) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER latest_signin;
-ALTER TABLE user_migrate CHANGE password_requested_at password_request datetime NULL;
-ALTER TABLE user_migrate MODIFY COLUMN salt varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL;
-ALTER TABLE user_migrate MODIFY COLUMN password varchar(255) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NOT NULL;
-ALTER TABLE user_migrate MODIFY COLUMN roles longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL;
-
--- Move fields to right place
+ALTER TABLE user_migrate CHANGE password_requested_at password_request datetime NULL AFTER token;
+ALTER TABLE user_migrate CHANGE roles roles longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci DEFAULT NULL AFTER password_request;
 ALTER TABLE user_migrate CHANGE username username varchar(255) AFTER roles;
-ALTER TABLE user_migrate CHANGE email email varchar(128) AFTER username;
-ALTER TABLE user_migrate CHANGE creation creation datetime AFTER email;
-ALTER TABLE user_migrate CHANGE enabled enabled tinyint(1) AFTER creation;
-ALTER TABLE user_migrate CHANGE salt salt varchar(255) AFTER enabled;
-ALTER TABLE user_migrate CHANGE password password varchar(255) AFTER salt;
-ALTER TABLE user_migrate CHANGE latest_signin latest_signin datetime AFTER password;
-ALTER TABLE user_migrate CHANGE password_request password_request datetime AFTER token;
-ALTER TABLE user_migrate CHANGE roles roles longtext AFTER password_request;
 
 -- Add missing fields
 -- In the following, un-comment to create the field if not already existing in your table. They are all required by c975L/UserBundle
--- ALTER TABLE user_migrate ADD identifier varchar(32) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER id;
 -- ALTER TABLE user_migrate ADD gender set('woman','man') CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER email;
 -- ALTER TABLE user_migrate ADD firstname varchar(48) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER gender;
 -- ALTER TABLE user_migrate ADD lastname varchar(48) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL AFTER firstname;
@@ -73,11 +73,11 @@ ALTER TABLE user_migrate CHANGE roles roles longtext AFTER password_request;
 -- Take care of the AFTER value, if you don't use all the fields
 -- ADDRESS
 /*
-ALTER TABLE user_migrate ADD address varchar(128) DEFAULT NULL AFTER roles;
+ALTER TABLE user_migrate ADD address varchar(128) DEFAULT NULL AFTER locale;
 ALTER TABLE user_migrate ADD address2 varchar(128) DEFAULT NULL AFTER address;
 ALTER TABLE user_migrate ADD postal varchar(10) DEFAULT NULL AFTER address2;
 ALTER TABLE user_migrate ADD town varchar(64) DEFAULT NULL AFTER postal;
-ALTER TABLE user_migrate ADD country varchar(64) DEFAULT NULL AFTER country;
+ALTER TABLE user_migrate ADD country varchar(64) DEFAULT NULL AFTER town;
 */
 -- BUSINESS
 /*
@@ -93,7 +93,7 @@ ALTER TABLE user_migrate ADD business_tva char(13) DEFAULT NULL AFTER business_s
 */
 -- SOCIAL
 /*
-ALTER TABLE user_migrate ADD social_network varchar(64) DEFAULT NULL AFTER business_tva;
+ALTER TABLE user_migrate ADD social_network varchar(24) DEFAULT NULL AFTER business_tva;
 ALTER TABLE user_migrate ADD social_id varchar(255) DEFAULT NULL AFTER social_network;
 ALTER TABLE user_migrate ADD social_token varchar(255) DEFAULT NULL AFTER social_id;
 ALTER TABLE user_migrate ADD social_picture varchar(255) DEFAULT NULL social_token;